--- client/reducers/reducer-trucksList.js
@@ -1,24 +1,0 @@
-export default function () {
-  return [
-    {
-      id: 1,
-      name: 'Curry Up Now',
-      handle: '@curryupnow',
-      lastTweet: {
-        timestamp: '2 hours ago',
-        message: 'closed on Saturdays'
-        },
-      pic: 'http://ww2.kqed.org/bayareabites/wp-content/uploads/sites/24/2011/02/curry-up-truck.jpg'
-    },
-    {
-      id: 2,
-      name: 'Senor Sisig',
-      handle: '@senorsisig',
-      lastTweet: {
-        timestamp: '20 minutes ago',
-        message: 'where am I?'
-        },
-      pic: 'http://insidescoopsf.sfgate.com/files/2015/10/Senor-Sisig-Truck.jpg'
-    }
-  ];
-};
