--- client/reducers/reducer-tweets.js
@@ -1,5 +1,0 @@
-
-
-export default function(){
-  return [];
-}
