+++ client/reducers/reducers.js
@@ -1,12 +1,10 @@
 import {combineReducers} from 'redux';
-import TweetReducer from './reducer_TweetList';
-import trucksList from './reducer-trucksList';
-import activeTruck from './reducer-activeTruck';
+import TruckReducer from './reducer_TruckList';
+// import trucksList from './reducer-trucksList';
+// import activeTruck from './reducer-activeTruck';
 
 const rootReducer = combineReducers({
-  tweet: TweetReducer,
-  trucks: trucksList,
-  activeTruck: activeTruck
+  trucks: TruckReducer
 });
 
 export default rootReducer;
